refactor: remove session controller and update imports to use localStorage utility
